requirements.txt
asgiref==3.8.1
attrs==24.2.0
Brotli==1.1.0
certifi==2024.8.30
charset-normalizer==3.3.2
click==8.1.7
defusedxml==0.7.1
dj-database-url==2.2.0
Django==5.1
django-filter==24.3
djangorestframework==3.15.2
drf-spectacular==0.27.2
drf-spectacular-sidecar==2024.7.1
gunicorn==23.0.0
h11==0.14.0
httpie==3.2.3
idna==3.8
inflection==0.5.1
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
markdown-it-py==3.0.0
mdurl==0.1.2
multidict==6.0.5
mysqlclient==2.2.4
packaging==24.1
pillow==10.4.0
psycopg2-binary==2.9.9
Pygments==2.18.0
PySocks==1.7.1
python-dotenv==1.0.1
PyYAML==6.0.2
referencing==0.35.1
requests==2.31.0
requests-toolbelt==1.0.0
rich==13.8.0
setuptools==74.0.0
sqlparse==0.5.1
typing_extensions==4.12.2
uritemplate==4.1.1
urllib3==2.2.2
uvicorn==0.30.6
whitenoise==6.7.0
python-decouple
django-allauth>=65.0.2
drf-extensions
django_cors_headers
PyPDF2
PyMuPDF
django-redis
dj-rest-auth[with_social]
djangorestframework-jwt
djangorestframework-simplejwt